Use FROM_EMAIL for web-components tests
