+++ django_app/frontend/package.json
@@ -58,6 +58,7 @@
     "@playwright/test": "^1.48.2",
     "@types/dompurify": "^3.0.5",
     "@types/showdown": "^2.0.6",
+    "dotenv": "^16.4.5",
     "parcel": "^2.12.0",
     "parcel-reporter-static-files-copy": "^1.5.3",
     "postcss": "^8.4.39",
