+++ django_app/frontend/tests-web-components/canned-prompts.spec.js
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-const signIn = require("./_sign-in.js");
+const { signIn } = require("./utils.js");
 
 test(`Clicking canned prompts updates the text input`, async ({ page }) => {
   await signIn(page);
